+++ src/core/songlist.ts
@@ -13,15 +13,6 @@ const LIST_LOAD_LIMIT = 30
 
 
 /**
- * 获取排序列表
- * @param source
- * @returns
- */
-export const getSortList = (source: LX.OnlineSource) => {
-  return songlistState.sortList[source]!
-}
-
-/**
  * 获取标签列表
  * @param source
  * @returns