+++ src/navigation/index.ts
@@ -31,7 +31,6 @@ const init = (callback: () => void | Promise<void>) => {
 }
 
 export * from './utils'
-export * from './event'
 export * from './hooks'
 
 export {