+++ src/navigation/utils.ts
@@ -1,6 +1,5 @@
 import { Navigation } from 'react-native-navigation'
 import {
-  VERSION_MODAL,
   PACT_MODAL,
 } from './screenNames'
 import themeState from '@/store/theme/state'
@@ -20,55 +19,6 @@ export const showPactModal = () => {
   void Navigation.showOverlay({
     component: {
       name: PACT_MODAL,
-      options: {
-        layout: {
-          componentBackgroundColor: 'transparent',
-        },
-        overlay: {
-          interceptTouchOutside: true,
-        },
-        statusBar: {
-          drawBehind: true,
-          visible: true,
-          style: getStatusBarStyle(theme.isDark),
-          backgroundColor: 'transparent',
-        },
-        navigationBar: {
-          // visible: false,
-          backgroundColor: theme['c-content-background'],
-        },
-        // animations: {
-
-        //   showModal: {
-        //     enter: {
-        //       enabled: true,
-        //       alpha: {
-        //         from: 0,
-        //         to: 1,
-        //         duration: 300,
-        //       },
-        //     },
-        //     exit: {
-        //       enabled: true,
-        //       alpha: {
-        //         from: 1,
-        //         to: 0,
-        //         duration: 300,
-        //       },
-        //     },
-        //   },
-        // },
-      },
-    },
-  })
-}
-
-export const showVersionModal = () => {
-  const theme = themeState.theme
-
-  void Navigation.showOverlay({
-    component: {
-      name: VERSION_MODAL,
       options: {
         layout: {
           componentBackgroundColor: 'transparent',