+++ src/store/version/index.ts
@@ -1,4 +1,0 @@
-export { default as versionState } from './state'
-export { default as versionAction } from './action'
-export * from './hook'
-