+++ src/utils/nativeModules/utils.ts
@@ -68,19 +68,6 @@ export const getWindowSize = async(): Promise<{ width: number, height: number }>
   return UtilsModule.getWindowSize()
 }
 
-export const onWindowSizeChange = (handler: (size: { width: number, height: number }) => void): () => void => {
-  UtilsModule.listenWindowSizeChanged()
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-  const eventEmitter = new NativeEventEmitter(UtilsModule)
-  const eventListener = eventEmitter.addListener('screen-size-changed', event => {
-    handler(event as { width: number, height: number })
-  })
-
-  return () => {
-    eventListener.remove()
-  }
-}
-
 export const isIgnoringBatteryOptimization = async(): Promise<boolean> => {
   return UtilsModule.isIgnoringBatteryOptimization()
 }