+++ src/utils/pixelRatio.ts
@@ -85,10 +85,6 @@ export const scaleSizeWR = (size: number) => {
   return size * 2 - scaleSizeW(size)
 }
 
-export const scaleSizeHR = (size: number) => {
-  return size * 2 - scaleSizeH(size)
-}
-
 export const scaleSizeAbsHR = (size: number) => {
   let scaleHeight = size * scale
   return size * 2 - Math.floor(scaleHeight / pixelRatio)